refactor: organize project into modules
