+++ index.html
@@ -51,6 +51,6 @@
   </div>
 
   <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
-  <script src="script.js"></script>
+  <script type="module" src="js/dashboard.js"></script>
 </body>
 </html>
