+++ lancamentos.html
@@ -156,6 +156,6 @@
     </div>
   </div>
 
-  <script src="script.js"></script>
+  <script type="module" src="js/lancamentos.js"></script>
 </body>
 </html>
