+++ relatorios.html
@@ -57,6 +57,6 @@
   </div>
 
   <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
-  <script src="script.js"></script>
+  <script type="module" src="js/relatorios.js"></script>
 </body>
 </html>
